+++ src/cli/commands/mc-download-file.ts
@@ -1,8 +1,14 @@
 import { CommanderStatic } from "commander";
 import { log } from "console";
+import * as fs from "fs";
+import * as path from "path";
+import * as util from "util";
+import { serviceCredentialLog } from "../../../dist/src/cli/commands/command-utils";
 import { MindSphereSdk } from "../../api/sdk/";
-import { decrypt, loadAuth } from "../../api/utils";
-import { errorLog, getColor } from "./command-utils";
+import { checksumFile, decrypt, loadAuth, retry, throwError } from "../../api/utils";
+import { errorLog, getColor, humanFileSize, verboseLog } from "./command-utils";
+import ora = require("ora");
+const streamPipeline = util.promisify(require("stream").pipeline);
 
 const color = getColor("magenta");
 
@@ -12,8 +18,6 @@ export default (program: CommanderStatic) => {
         .alias("df")
         .option("-f, --file <fileToDownload>", "file to download from the file service")
         .option("-h, --filepath [filepath]", "file path in the mindsphere", "")
-        .option("-y, --retry <number>", "retry attempts before giving up", 3)
-        .option("-l, --parallel <number>", "parallel chunked downloads", 3)
         .option("-i, --assetid <assetid>", "asset id from the mindsphere")
         .option("-p, --passkey <passkey>", `passkey`)
         .option("-y, --retry <number>", "retry attempts before giving up", 3)
@@ -31,9 +35,48 @@ export default (program: CommanderStatic) => {
                         basicAuth: decrypt(auth, options.passkey),
                         tenant: auth.tenant
                     });
-                    const iotFile = sdk.GetIoTFileClient();
-                    const download = (await iotFile.GetFile(options.assetid, options.filepath));
-                    
+                    const iotFileClient = sdk.GetIoTFileClient();
+                    let fullpath = options.filepath ? `${options.filepath}/${options.file}` : `${options.file}`;
+                    fullpath = fullpath.replace("//", "/");
+
+                    let filter = `name eq ${path.basename(fullpath)}`;
+                    if (path.dirname(fullpath) !== ".") filter += ` and path eq ${path.dirname(fullpath)}/`;
+
+                    const fileInfo = await iotFileClient.GetFiles(options.assetid, {
+                        filter: filter
+                    });
+
+                    fileInfo.length !== 1 &&
+                        errorLog(`There were ${color(fileInfo.length)} files found with that name and path.`, true);
+
+                    const spinner = ora("downloading file");
+                    !options.verbose && spinner.start();
+
+                    verboseLog(JSON.stringify(fileInfo, null, 2), options.verbose);
+                    verboseLog(
+                        `Downloading file ${color(fileInfo[0].name)} with size of ${color(
+                            humanFileSize(fileInfo[0].size || 0)
+                        )} from MindSphere.`,
+                        options.verbose,
+                        spinner
+                    );
+
+                    const startDate = new Date();
+
+                    const download = await retry(options.retry, () => iotFileClient.GetFile(options.assetid, fullpath));
+                    !download.ok && throwError(`Unexpected response ${download.statusText}`);
+                    const outputfile = fileInfo[0].name || "output.bin";
+                    const file = fs.createWriteStream(outputfile);
+                    await streamPipeline(download.body, file);
+
+                    const endDate = new Date();
+
+                    !options.verbose && spinner.succeed("Done");
+
+                    const hash = await checksumFile("md5", outputfile);
+
+                    log(`Download time: ${(endDate.getTime() - startDate.getTime()) / 1000} seconds`);
+                    log(`\nYour file ${color(outputfile)} with md5 hash ${hash} was successfully downloaded`);
                 } catch (err) {
                     errorLog(err, options.verbose);
                 }
@@ -42,11 +85,21 @@ export default (program: CommanderStatic) => {
         .on("--help", () => {
             log("\n  Examples:\n");
             log(
-                `    mc download-file -f CHANGELOG.md  --assetid 5..f  \t\t\t\t download file CHANGELOG.md from specified asset`
+                `    mc download-file -f CHANGELOG.md  --assetid 5..f  \t\t\t\t download file ${color(
+                    "CHANGELOG.md"
+                )} from specified asset`
             );
             log(
-                `    mc download-file --file  CHANGELOG.md  --assetid 5...f --filepath upload \t download file upload/CHANGELOG.md from specified asset`
+                `    mc download-file --file  CHANGELOG.md  --assetid 5...f --filepath upload \t download file ${color(
+                    "upload/CHANGELOG.md"
+                )} from specified asset`
             );
+            log(
+                `    mc download-file --file  upload/CHANGELOG.md  --assetid 5...f \t\t download file ${color(
+                    "upload/CHANGELOG.md"
+                )} from specified asset`
+            );
+            serviceCredentialLog();
         });
 };